+++ plugins/self-service/src/components/RouteView/RouteView.tsx
@@ -77,7 +77,11 @@ const RouteViewContent = () => {
         <Route path="create">
           <Route
             path="templates/:namespace/:templateName"
-            element={<CreateTask />}
+            element={
+              <RequirePermission permission={templatesViewPermission}>
+                <CreateTask />
+              </RequirePermission>
+            }
           />
           <Route
             path="tasks"
